+++ Source/Draw/Camera/CameraControl.cpp
@@ -85,7 +85,7 @@ void CameraControl::MakeCallback() {
 			}
 
 			if (Engine::Callback::pressTap(Engine::VirtualTap::MIDDLE)) {
-				Move(Engine::Callback::deltaMousePos() * Engine::Core::deltaTime());
+				Move(Engine::Callback::deltaMousePos() * (float)Engine::Core::deltaTime());
 			}
 		}
 	);
